Add a pending test regarding dependencies of macro expansions
All symbols looked up during the expansion of a macro should be
registered as dependencies of the macro client, which is not the
case at the moment.

The idea is to add a compiler plugin to the sbt project that attaches
the symbols that have been looked up during the expansion of a macro,
so that we can register the dependencies between the macro client and
these symbols.