Add 3 pending tests about dependencies of macro expansions
All symbols looked up during the expansion of a macro should be
registered as dependencies of the macro client, which is not the
case at the moment.

The idea is to add a compiler plugin to the sbt project that attaches
the symbols that have been looked up during the expansion of a macro,
so that we can register the dependencies between the macro client and
these symbols.

The first test makes sure that whenever a method is added to a class
that is inspected by a macro, then the clients of this macro will be
recompiled.

The second test tests that whenever a method is removed from a class
that is inspected by a macro, then the clients of this macro will be
recompiled.

The third test checks that renaming the arguments of a method from
a class that is inspected by a macro will also trigger the
recompilation of the clients of this macro.