Extract and register symbols reported by scalahost
If a project is compiled using the scalahost compiler plugin, the
plugin will attach the symbols that have been touched to the
expansions of macros. This commit adds the extraction and registration
of these symbols, so that new dependency relations are created.

Imagine that a macro lists the members of a certain class. Prior to
this commit, adding a new method to this class would not have
triggered the recompilation of the clients of this macro. Now, it does.