Invalidate macros whose internal deps are changed
Whenever any transitive dependency of a macro provider is changed, then
all macro clients should be recompiled.

This commit introduces a new kind of dependency
(`DependencyFromMacroImpl`)  along with the
(approximate) extraction logic for it.

The invalidation logic has been adapted so that is a file on which a
macro provider depends, then the macro provider is recompiled.

Marks the scripted test `source-dependencies/macro-graph-simple` has
passing.