Export max depth for transitive macros deps to option
This commit introduces a new field in `IncOptions` that allows to define
the maximum depth at which we want to look for recompiled dependencies
of macro providers. The default depth is set to 5.

The value of this field can be changed in a project's configuration
using:

    incOptions := incOptions.value.withMacroTransitiveMaxDepth(5)

When this field is set to zero, the incremental compiler will only
check, for each macro provider, whether its immediate dependencies have
been recompiled.