Better resolution of compilation times
The previous implementation lead to many re-computations of the same
information over and over.

This new implementation caches the result, so that no file is checked
more than once for recompilation. The result is that we do not need any
longer an option to limit the depth of the exploration that we want to
perform.

The setting that has been introduced in the previous commit has been
removed in favor of a binary flag that controls whether or not we want
to enable invalidation of macro providers based on the recompilation of
their transitive dependencies.