Tentative fix for high memory consumption
From the reproduction case, I gathered that the issue might come from
the very long macro expansions chain that is involved in the generation
of picklers and unpicklers.

My idea to fix this issue is to consider only dependencies from a tree
to a symbol if and only if at least one of them is not synthetic.

A tree is considered synthetic iff its position if not a range (this
seems to be the criterion used in scalac).