Diagnose problems with Travis
For some time now we've been suspecting that we get spurious test
failures because of the resolution of `File.lastModified`.

The problem could come from sbt not seeing changes to sources or
products because they've been changed too quickly between 2 compilation
runs.

I'm trying to always depend on hashes to tell whether a file has changed.